pyproject.toml
[tool.poetry]
name = "datadog-api-client-generator"
version = "0.1.0"
description = "Tooling for generating Datadog API clients"
authors = ["Datadog <support@datadoghq.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.6"
PyYAML = "^6.0"
jsonref = "^1.1.0"
click = "^8.1.7"
jinja2 = "^3.1.2"
m2r2 = "0.3.3.post2"

[tool.poetry.scripts]
generate = "datadog_api_client_generator.cli:cli"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
ruff = "^0.9.9"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D"  # pydocstyle
]
ignore = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
